+++ server/game/GameEngine.hpp
@@ -60,9 +60,11 @@ public:
      * The entity is not yet spawned.  The gateway should forward the client's
      * JOIN message to handlePlayerInput(), which will spawn the entity with
      * the requested EntityType and send the initial snapshot.
+     *
+     * @param sx/sy/sz Spawn position in sub-voxels (1 voxel = SUBVOXEL_SIZE units).
      */
     void queuePendingPlayer(GatewayId gwId, PlayerId playerId,
-                            float sx, float sy, float sz);
+                            int32_t sx, int32_t sy, int32_t sz);
 
     /**
      * @brief Spawn a new player entity with the given type.
@@ -72,22 +74,24 @@ public:
      *
      * @param gwId     Gateway the player connected through.
      * @param playerId Persistent player identifier.
-     * @param sx/sy/sz Spawn world coordinates (voxels).
+     * @param sx/sy/sz Spawn position in sub-voxels (1 voxel = SUBVOXEL_SIZE units).
      * @param type     Entity type (selects physics mode).
      */
     void addPlayer(GatewayId gwId, PlayerId playerId,
-                   float sx, float sy, float sz,
+                   int32_t sx, int32_t sy, int32_t sz,
                    EntityType type = EntityType::GHOST_PLAYER);
 
     /** @brief Destroy a player entity and remove it from all chunk sets. */
     void removePlayer(PlayerId playerId);
 
     /**
-     * @brief Directly set a player's world position (voxel coordinates).
+     * @brief Directly set a player's world position (sub-voxel coordinates).
      * Intended for test setup and administrative use.
      * No-op if the player does not exist.
+     *
+     * @param sx/sy/sz Position in sub-voxels (1 voxel = SUBVOXEL_SIZE units).
      */
-    void teleportPlayer(PlayerId playerId, float sx, float sy, float sz);
+    void teleportPlayer(PlayerId playerId, int32_t sx, int32_t sy, int32_t sz);
 
     // ── Main loop ─────────────────────────────────────────────────────────
 
@@ -146,7 +150,7 @@ private:
     std::unordered_map<GatewayId, GatewayInfo>             gateways;
     std::unordered_map<PlayerId,  entt::entity>            playerEntities;
 
-    struct PendingPlayer { GatewayId gwId; float sx, sy, sz; };
+    struct PendingPlayer { GatewayId gwId; int32_t sx, sy, sz; };
     std::unordered_map<PlayerId, PendingPlayer>            pendingPlayers;
 
     int32_t  tickCount{0};
